speeded up clipboard transfer by avoiding a selection request when it wasn't necessary.  (in particular, we were getting the clipboard update time from the owner then emptying the clipboard, so we didn't need to get the time.  worse, most owners don't support getting the time and we often timed out.)
also fixed a multithread bug using the X display.  we were using
a CThread to send an event after a timeout while we were waiting
in XIfEvent().  this necessarily involved two threads calling
into Xlib at once, which is not allowed.  now using polling to
do the timeout because Xlib doesn't have a function to get
events with a timeout.