Added support on OS X for screensaver synchronization, sleeping and fast user switching.  OS X Server also now captures global hot keys (cmd+tab, F9, etc.) and sends them to the client.  Changes mostly due to lorenz schori.  Some bug fixes (in lorenz's code and in synergy) and integration into automake by crs (had to add support for Objective C++ files) and the XCode project.
This change also adds flags to events.  Flags can cause events
to be dispatched immediately and to not free event data.  Both
features are used in the new code.

This change adds events on IScreen for notification of the
system going to sleep or waking up (or the user's session
being somehow suspended and restored).  CClient and synergys
now listen for and respond to these events.  CMSWindowsScreen
used to use IJobs for handling these events synchronously.  It
now uses the new IScreen events and delivers them immediately.